+++ coner/src/pages/Request/RequestModifyPage.jsx
@@ -9,7 +9,7 @@ const RequestModifyPage = () => {
 
   return (
     <Container>
-      <NavHeader to="/request/step1" title="내 정보 수정" />
+      <NavHeader to="/request/step5" title="내 정보 수정" />
       <Modify />
     </Container>
   );
