+++ coner/src/pages/partner/Steps/PartnerStep5.jsx
@@ -1,66 +1,171 @@
 import React, { useState, useEffect } from "react";
 import styled from "styled-components";
 import StepHeader from "../../../components/common/Header/StepHeader";
-import { useNavigate, useParams } from "react-router-dom";
+import TextField from "../../../components/ui/formControls/TextField";
+import StepSegmentedToggle from "../../../components/ui/StepSegementedToggle";
 import Button from "../../../components/ui/Button";
+import AddressModal, {
+  SERVICE_AREAS,
+} from "../../../components/common/Modal/AddressModal";
+import Modal from "../../../components/common/Modal/Modal";
+import {
+  useNavigate,
+  useLocation,
+  useSearchParams,
+  useParams,
+} from "react-router-dom";
 import { useRequest } from "../../../context/context";
 import { useFunnelStep } from "../../../analytics/useFunnelStep";
+import { useAuth } from "../../../context/AuthProvider";
 
 const PartnerStep5 = () => {
   const navigate = useNavigate();
   const { partnerId } = useParams();
-  const { requestData, updateRequestData } = useRequest();
+  const location = useLocation();
+  const [searchParams] = useSearchParams();
+  const { requestData, updateRequestData, updateRequestDataMany } =
+    useRequest();
+  const { currentUser, userInfo } = useAuth();
+  const isLoggedIn = !!currentUser;
+  const isReadOnly = isLoggedIn && !!userInfo;
 
+  // 퍼널: 5단계
   const { onAdvance } = useFunnelStep({ step: 5 });
 
-  const [additionalRequest, setAdditionalRequest] = useState("");
+  const [phone, setPhone] = useState(requestData.customer_phone || "");
+  const [address, setAddress] = useState(requestData.customer_address || "");
+  const [detailaddress, setDetailaddress] = useState(
+    requestData.customer_address_detail || ""
+  );
+  const [customerType, setCustomerType] = useState(
+    requestData.customer_type || ""
+  );
+  const [isAddressOpen, setIsAddressOpen] = useState(false);
 
-  const handleTextChange = (e) => {
-    const value = e.target.value;
-    setAdditionalRequest(value);
-  };
+  // 로그인 사용자 정보 자동 입력
+  useEffect(() => {
+    if (!userInfo || !isReadOnly) return;
+
+    const formattedPhone = formatPhoneNumber(userInfo.phone || "");
+    setPhone(formattedPhone);
+    setAddress(userInfo.address || "");
+    setDetailaddress(userInfo.address_detail || "");
+    setCustomerType(userInfo.job || "");
+
+    updateRequestDataMany({
+      customer_phone: formattedPhone,
+      customer_address: userInfo.address || "",
+      customer_address_detail: userInfo.address_detail || "",
+      customer_type: userInfo.job || "",
+    });
+  }, [userInfo, isReadOnly]);
+
+  useEffect(() => {
+    const restoredService =
+      location.state?.selectedService || searchParams.get("service_type");
+    if (restoredService && !requestData.service_type) {
+      updateRequestData("service_type", restoredService);
+    }
+  }, [
+    location.state,
+    searchParams,
+    requestData.service_type,
+    updateRequestData,
+  ]);
+
   const handleHelpClick = () => {
     window.open("http://pf.kakao.com/_jyhxmn/chat");
   };
 
-  useEffect(() => {
-    if (requestData.detailInfo) {
-      const lines = requestData.detailInfo.split("\n");
-      const additionalLines = lines.filter(
-        (line) =>
-          !line.includes("중고에어컨으로 원해요") &&
-          !line.includes("신규에어컨으로 원해요")
-      );
-      setAdditionalRequest(additionalLines.join("\n"));
-    }
-  }, [requestData.detailInfo]);
-
   const handleNext = () => {
-    const existingDetail = requestData.detailInfo || "";
+    if (!phone || !address || !detailaddress || !customerType) {
+      alert("모든 정보를 입력해주세요.");
+      return;
+    }
 
-    const purchaseTypeLines = existingDetail
-      .split("\n")
-      .filter(
-        (line) =>
-          line.includes("중고에어컨으로 원해요") ||
-          line.includes("신규에어컨으로 원해요")
-      );
+    const phoneNumbers = phone.replace(/[^\d]/g, "");
+    if (phoneNumbers.length !== 11) {
+      alert("올바른 휴대폰 번호를 입력해주세요.");
+      return;
+    }
 
-    const additionalText = additionalRequest.trim();
-    const newDetail =
-      purchaseTypeLines.length > 0
-        ? additionalText
-          ? `${purchaseTypeLines.join("\n")}\n${additionalText}`
-          : purchaseTypeLines.join("\n")
-        : additionalText;
-
-    updateRequestData("detailInfo", newDetail);
+    updateRequestDataMany({
+      customer_phone: phoneNumbers,
+      customer_address: address,
+      customer_address_detail: detailaddress,
+      customer_type: customerType,
+    });
 
     onAdvance(6);
     navigate(`/partner/step6/${partnerId}`);
   };
 
-  const currentStep = additionalRequest.trim() ? 9 : 8;
+  // 휴대폰 번호 포맷팅 (000-0000-0000)
+  const formatPhoneNumber = (value) => {
+    const numbers = value.replace(/[^\d]/g, "");
+    if (numbers.length <= 3) return numbers;
+    if (numbers.length <= 7)
+      return `${numbers.slice(0, 3)}-${numbers.slice(3)}`;
+    return `${numbers.slice(0, 3)}-${numbers.slice(3, 7)}-${numbers.slice(
+      7,
+      11
+    )}`;
+  };
+
+  const handlePhoneChange = (e) => {
+    if (isReadOnly) return;
+    const value = e.target ? e.target.value : e;
+    const formatted = formatPhoneNumber(value);
+    setPhone(formatted);
+    const phoneNumbers = formatted.replace(/[^\d]/g, "");
+    updateRequestData("customer_phone", phoneNumbers);
+  };
+
+  const handleDetailAddressChange = (e) => {
+    if (isReadOnly) return;
+    const value = e.target ? e.target.value : e;
+    setDetailaddress(value);
+    updateRequestData("customer_address_detail", value);
+  };
+
+  const handleAddressClick = () => {
+    if (!isReadOnly) {
+      setIsAddressOpen(true);
+    }
+  };
+
+  const handleAddressSelect = (addr) => {
+    setAddress(addr);
+    updateRequestData("customer_address", addr);
+    setIsAddressOpen(false);
+  };
+
+  const handleCustomerTypeChange = (value) => {
+    if (!isReadOnly) {
+      setCustomerType(value);
+      updateRequestData("customer_type", value);
+    }
+  };
+
+  const phoneNumbers = phone.replace(/[^\d]/g, "");
+  const isPhoneComplete = phoneNumbers.length === 11;
+  const isAddressComplete =
+    address.trim().length > 0 && detailaddress.trim().length > 0;
+  const isCustomerTypeComplete = customerType.trim().length > 0;
+
+  let currentStep = 6;
+  let title = "휴대폰 번호를 입력해주세요.";
+
+  if (isPhoneComplete && !isAddressComplete) {
+    currentStep = 7;
+    title = "주소를 입력해주세요.";
+  } else if (isPhoneComplete && isAddressComplete && !isCustomerTypeComplete) {
+    currentStep = 8;
+    title = "개인이신가요? 사업이신가요?";
+  } else if (isPhoneComplete && isAddressComplete && isCustomerTypeComplete) {
+    currentStep = 9;
+    title = "확인버튼을 눌러주세요";
+  }
 
   return (
     <PageContainer>
@@ -71,42 +176,117 @@ const PartnerStep5 = () => {
           totalSteps={10}
         />
         <ContentSection>
-          <PageTitle>추가 요청사항이 있으신가요?</PageTitle>
+          <PageTitle>{title}</PageTitle>
 
+          {/* 로그인 사용자 정보 수정 링크 */}
+          {isReadOnly && (
+            <ModifyLink
+              onClick={() =>
+                navigate(`/partner/modify/${partnerId}`, {
+                  state: { from: "partnerstep5" },
+                })
+              }
+            >
+              내 정보 (주소 / 고객유형) 수정하러가기
+            </ModifyLink>
+          )}
+
+          {/* 3단계: 의뢰인 유형 */}
+          {currentStep >= 8 && (
+            <FormGroup>
+              <Label>의뢰인 유형</Label>
+              <StepSegmentedToggle
+                value={customerType}
+                onChange={handleCustomerTypeChange}
+                options={[
+                  { label: "사업장(기업/매장)", value: "사업장(기업/매장)" },
+                  { label: "개인(가정)", value: "개인(가정)" },
+                ]}
+                disabled={isReadOnly}
+              />
+            </FormGroup>
+          )}
+
+          {/* 2단계: 주소 */}
+          {currentStep >= 7 && (
+            <FormGroup>
+              <ClickableTextField
+                label="주소"
+                size="stepsize"
+                value={address}
+                placeholder="주소를 입력해주세요"
+                onClick={handleAddressClick}
+                readOnly
+              />
+              <p style={{ marginBottom: "5px" }}></p>
+              <TextField
+                size="stepsize"
+                value={detailaddress}
+                onChange={handleDetailAddressChange}
+                placeholder="상세주소를 입력해주세요"
+                readOnly={isReadOnly}
+              />
+              {/* 서울 지역 제한 안내 */}
+              <HelperText>서울 지역으로만 제한되어 있습니다.</HelperText>
+            </FormGroup>
+          )}
+
+          {/* 1단계: 휴대폰 번호 (항상 표시) */}
           <FormGroup>
-            <Label>추가 요청사항</Label>
-            <TextArea
-              value={additionalRequest}
-              onChange={handleTextChange}
-              placeholder="추가 요청사항을 입력해주세요 (선택사항)"
-              rows={6}
+            <TextField
+              label="휴대폰 번호"
+              size="stepsize"
+              value={phone}
+              onChange={handlePhoneChange}
+              placeholder="000-0000-0000"
+              maxLength={13}
+              readOnly={isReadOnly}
             />
           </FormGroup>
         </ContentSection>
       </ScrollableContent>
 
-      <FixedButtonArea>
-        <Button fullWidth size="stepsize" onClick={handleNext}>
-          확인
-        </Button>
-        <HelpButton onClick={handleHelpClick}>
-          <HelpText>도움이 필요해요</HelpText>
-          <svg
-            xmlns="http://www.w3.org/2000/svg"
-            width="6"
-            height="10"
-            viewBox="0 0 6 10"
-            fill="none"
-          >
-            <path
-              d="M1 1L5 5L1 9"
-              stroke="#A0A0A0"
-              strokeLinecap="round"
-              strokeLinejoin="round"
+      {/* 하단 고정 버튼 - 조건부 렌더링 */}
+      {currentStep === 9 && (
+        <FixedButtonArea>
+          <Button fullWidth size="stepsize" onClick={handleNext}>
+            확인
+          </Button>
+          <CSButtonContainer>
+            <CSButton onClick={handleHelpClick}>
+              <CSButtonText>도움이 필요해요</CSButtonText>
+              <svg
+                xmlns="http://www.w3.org/2000/svg"
+                width="8"
+                height="14"
+                viewBox="0 0 8 14"
+                fill="none"
+              >
+                <path d="M0.999999 13L7 7L1 1" stroke="#A0A0A0" />
+              </svg>
+            </CSButton>
+          </CSButtonContainer>
+        </FixedButtonArea>
+      )}
+
+      {/* 주소 검색 모달 */}
+      {isAddressOpen && (
+        <Modal
+          open={isAddressOpen}
+          onClose={() => setIsAddressOpen(false)}
+          title="주소 검색"
+          width={420}
+          containerId="rightbox-modal-root"
+        >
+          <div style={{ width: "100%", height: "70vh" }}>
+            <AddressModal
+              onSelect={handleAddressSelect}
+              onClose={() => setIsAddressOpen(false)}
+              serviceAreas={SERVICE_AREAS}
             />
-          </svg>
-        </HelpButton>
-      </FixedButtonArea>
+          </div>
+        </Modal>
+      )}
     </PageContainer>
   );
 };
@@ -165,81 +345,61 @@ const PageTitle = styled.h1`
   }
 `;
 
-const FormGroup = styled.div`
-  margin-bottom: 24px;
-`;
-
 const Label = styled.p`
-  margin-bottom: 8px;
-  font-size: ${({ theme }) => theme.font.size.body};
+  margin-bottom: 6px;
+  font-size: ${({ theme }) => theme.font.size.bodySmall};
   font-weight: ${({ theme }) => theme.font.weight.medium};
   color: ${({ theme }) => theme.colors.subtext};
 `;
 
-const TextArea = styled.textarea`
-  width: 100%;
-  min-height: 120px;
-  padding: 16px;
-  border: 1px solid #d6d6d6;
-  border-radius: 8px;
-  font-size: ${({ theme }) => theme.font.size.bodyLarge};
-  font-family: inherit;
-  color: ${({ theme }) => theme.colors.text};
-  background: ${({ theme }) => theme.colors.bg};
-  resize: vertical;
-  line-height: 1.5;
+const FormGroup = styled.div`
+  margin-bottom: 24px;
+`;
 
-  @media (max-width: ${({ theme }) => theme.font.breakpoints.mobile}) {
-    min-height: 100px;
-  }
-
-  &::placeholder {
-    color: #999;
-  }
-
-  &:focus {
-    outline: none;
-    border-color: #007bff;
-    box-shadow: 0 0 0 2px rgba(0, 123, 255, 0.25);
-  }
-
-  &::-webkit-scrollbar {
-    width: 6px;
-  }
-
-  &::-webkit-scrollbar-track {
-    background: #f1f1f1;
-    border-radius: 3px;
-  }
-
-  &::-webkit-scrollbar-thumb {
-    background: #c1c1c1;
-    border-radius: 3px;
-  }
-
-  &::-webkit-scrollbar-thumb:hover {
-    background: #a8a8a8;
+const ClickableTextField = styled(TextField)`
+  cursor: pointer;
+  input {
+    cursor: pointer;
   }
 `;
 
-const HelpButton = styled.button`
+const HelperText = styled.p`
+  color: ${({ theme }) => theme.colors.subtext};
+  font-weight: ${({ theme }) => theme.font.weight.regular};
+  font-size: ${({ theme }) => theme.font.size.bodySmall};
+  padding: 5px 0 0 0;
+  margin: 0;
+`;
+
+const ModifyLink = styled.a`
+  font-size: ${({ theme }) => theme.font.size.bodySmall};
+  font-weight: ${({ theme }) => theme.font.weight.bold};
+  text-decoration: underline;
+  color: ${({ theme }) => theme.colors.subtext};
+  padding: 10px;
+  cursor: pointer;
+  display: block;
+  margin-bottom: 20px;
+`;
+
+const CSButtonContainer = styled.div`
+  display: flex;
+  justify-content: center;
+  margin-top: 20px;
+`;
+
+const CSButton = styled.button`
+  color: ${({ theme }) => theme.colors.text};
   display: flex;
   align-items: center;
-  justify-content: center;
-  gap: 6px;
+  gap: 8px;
   background: none;
   border: none;
   cursor: pointer;
-  margin: 20px auto 0 auto;
-  padding: 8px;
-
-  &:hover {
-    background-color: #f8f9fa;
-    border-radius: 4px;
-  }
 `;
 
-const HelpText = styled.span`
+const CSButtonText = styled.p`
+  margin: 0;
   font-size: ${({ theme }) => theme.font.size.bodyLarge};
   color: #a0a0a0;
 `;
